// lib/Q/sys/Corelib.d.ts
export declare type GFunction = Function | reflection.GenericType | reflection.DelayedType;
// ...
export declare namespace string {
    function IsPrintable(keyCode: number, charCode: number): boolean;
}
export declare namespace helper {
    function TryCatch<T>(owner: any, Try: (...args: any[]) => T, Catch?: (e: Error, ...args: any[]) => T, params?: any[]): T;
}
export declare namespace basic {
// ...
    interface IEventHandler extends IDisposable {
        Started: boolean;
        Start(): any;
        Pause(): any;
        Dispose(): any;
        Reset(): any;
    }
// ...
    interface IJob {
        Name: string;
        Todo?(job: bind.JobInstance, e: bind.EventArgs<any, any>): void;
        Check?(job: bind.JobInstance, e: bind.EventArgs<any, any>): void;
        OnError?(job: bind.JobInstance, e: bind.EventArgs<any, any>): void;
        OnInitialize?(job: bind.JobInstance, e: bind.EventArgs<any, any>): void;
        OnScopDisposing?(job: bind.JobInstance, e: bind.EventArgs<any, any>): void;
    }
// ...
    interface IId {
        Id: number;
    }
// ...
}
// ...
export declare namespace reflection {
    interface ICallHistory {
        caller: any;
        arguments: any[];
        fn: Function;
    }
    type Method<RET, T extends (...args: any[]) => RET> = MethodGroup<RET, T> | T | basic.ITBindable<T>;
    class MethodGroup<RET, T extends (...args: any[]) => RET> implements basic.ITBindable<T> {
        Owner: any;
        private _list;
        constructor(f?: Method<RET, T>, Owner?: any);
        Invoke: T;
        add(m: T | Method<RET, T>): this;
        With(owner: any, ...args: any[]): RET;
        Clone(): MethodGroup<RET, T>;
    }
    function ToMethodGroup<RET, T extends (...args: any[]) => RET>(x: Method<RET, T>): MethodGroup<{}, T>;
    function Invoke<RET, T extends (...args: any[]) => RET>(f: Method<RET, T>, owner: any, args: any[]): RET;
    interface IDebuggerInfo {
        obsArgs: boolean;
        Stack?: ICallHistory[];
        debug?: boolean;
        save?: boolean;
        callback?: Function;
        fn: Function;
        proxy?: Function;
        ReCalc?: (callHistory: ICallHistory | number, direct: boolean) => any;
    }
    function isInstanceOfClassName(instance: any, className: any): boolean;
    function isInstanceOfClass(instance: any, type: any): boolean;
    function _isInstanceOf(type: Function, superType: Function): boolean;
    function GetBaseType(type: any): any;
    function GetBaseTypes(type: any, totype?: any): typeof Object[];
    function IsInstanceOf(type: any, superType: any): boolean;
    class Type {
        private passed;
        type: Function;
        constructor(type: any);
        _getPath(root: any): any;
        GetType(root: any): any;
    }
    class GenericType {
        Constructor: Function;
        Params: Function[];
        prototype: any;
        constructor(Constructor: Function, Params: Function[], base: Function);
        readonly base: Function;
        GetBaseType(): Function;
        static GetType(type: Function, params?: Function[], checkOnly?: boolean, base?: Function): GenericType | Function;
        private static i(f);
        static IsInstanceOf(type: any, superType: any): any;
        static _isInstanceOf: (((type: Function, superType: Function) => boolean) | ((type: Function, superGType: GenericType) => boolean) | ((gtype: GenericType, superGType: GenericType) => boolean) | ((gtype: GenericType, superType: Function) => boolean))[];
    }
    class DelayedType {
        readonly Type: Function;
        private _type;
        constructor(type: () => Function);
    }
    namespace Observable {
        function observeProperty(obj: any, propName: string, evnt: string): void;
        function setObservableProperty<T>(obj: any, propName: string, get: () => T, set: (val: T) => void, evnt: string): void;
        function ObjectToObservable(o: Object): void;
    }
    function IsClass(obj: ObjectConstructor): boolean;
    function IsPrototype(obj: any): boolean;
    function IsInstance(obj: any): boolean;
}
// ...
export declare namespace bind {
    class Job implements basic.IJob {
        Name: string;
        Todo: (job: JobInstance, e: bind.EventArgs<any, any>) => void;
        Check: (job: JobInstance, e: bind.EventArgs<any, any>) => void;
        OnError: (job: JobInstance, e: bind.EventArgs<any, any>) => void;
        OnInitialize: (job: JobInstance, e: bind.EventArgs<any, any>) => void;
        OnScopDisposing: (job: JobInstance, e: bind.EventArgs<any, any>) => void;
        constructor(Name: string, Todo: (job: JobInstance, e: bind.EventArgs<any, any>) => void, Check?: (job: JobInstance, e: bind.EventArgs<any, any>) => void, OnError?: (job: JobInstance, e: bind.EventArgs<any, any>) => void, OnInitialize?: (job: JobInstance, e: bind.EventArgs<any, any>) => void, OnScopDisposing?: (job: JobInstance, e: bind.EventArgs<any, any>) => void);
    }
    class Jobs implements basic.IJob {
        Name: string;
        Todo(job: JobInstance, e: bind.EventArgs<any, any>): void;
        Check(job: JobInstance, e: bind.EventArgs<any, any>): void;
        OnError(job: JobInstance, e: bind.EventArgs<any, any>): void;
        OnInitialize(job: JobInstance, e: bind.EventArgs<any, any>): void;
        OnScopDisposing(job: JobInstance, e: bind.EventArgs<any, any>): void;
        constructor(Name: string);
        push(jobName: any): void;
    }
    class JobInstance implements EventListenerObject {
        Scop: bind.Scop;
        job: basic.IJob;
        dom: Node;
        Control: UI.JControl;
        private _events;
        Handle: (ji: JobInstance, e: Event) => void;
        addEventListener(name: string, event: string, delegate: EventListenerOrEventListenerObject | any): void;
        removeEventListener(name: string): void;
        getEvent(name: string): EventListenerOrEventListenerObject;
        constructor(Scop: bind.Scop, job: basic.IJob, dom: Node);
        private propb;
        private ValueChanged(s, e);
        Dispose(): void;
        IsDisposed: any;
        _store: any;
        addValue(name: string, value: any): void;
        getValue(name: string): any;
        Checker: (value: any) => boolean;
        Ischanging: boolean;
        handleEvent(e: Event): void;
    }
    function GetJob(name: string): basic.IJob;
    function Register(job: basic.IJob, override?: boolean): basic.IJob;
}
export declare namespace thread {
    interface IDispatcherCallback {
        callback: (delegate: (...param: any[]) => void, param: any, _this: any) => void;
        params: JobParam;
        _this: any;
        optimizable: boolean;
        isWaiting: boolean;
        id: number;
        children: IDispatcherCallback[];
        ce: number;
    }
    class JobParam {
        id: number;
        params: any[];
        constructor(id: number, params?: any[]);
        Set(...params: any[]): this;
        Set1(params: any[]): this;
        Clone(): JobParam;
    }
    class Dispatcher {
        static OnIdle(owner: any, callback: () => void, once?: boolean): void;
        static InIdle(): boolean;
        static GC(): void;
        static clone(ojob: IDispatcherCallback, params: any[], __this?: any): IDispatcherCallback;
        static cretaeJob(delegate: (...param: any[]) => void, param: any[], _this: any, optimizable: boolean): JobParam;
        static Clear(o: JobParam): void;
        static readonly CurrentJob: IDispatcherCallback;
        private static start();
        static Push(ojob: JobParam, params?: any[], _this?: any): IDispatcherCallback;
        static call(_this: any, fn: Function, ...args: any[]): void;
        static IsRunning(): boolean;
    }
}
export declare namespace bind {
    class DProperty<T, P> {
        Attribute: PropertyAttribute;
        Name: string;
        Type: GFunction;
// ...
    }
    class EventArgs<T, P> implements basic.IDisposable {
        static New<T, P>(prop: DProperty<T, P>, ithis: P, _old: T, _new: T): EventArgs<T, P>;
        prop: DProperty<T, P>;
        __this: P;
        _old: T;
        _new: T;
        IsValid: boolean;
        Dispose(): void;
    }
    class Ref<T> {
        private _key;
        key: T;
    }
    class EventListener<T extends Function> implements basic.IDisposable {
        private _deleagtes;
        private key;
        private isSingliton;
        constructor(key: any, isSingliton?: boolean);
        On: T;
        private locks;
        Off: T;
        private lock;
        Invoke(key: Object, params: any[]): void;
        Invok(key: Object, callBack: (delegate: T) => any): any;
        PInvok(key: Object, params: any[], owner?: any): any;
        Add(delegate: T, key?: any): void;
        Remove(key: any): void;
        private _store;
        Dispose(): void;
    }
// ...
    class Controller extends DObject implements basic.IDisposable {
        OnNodeLoaded(): any;
        readonly MainControll: UI.JControl;
        static Attach(control: UI.JControl, data?: any | Scop): Controller;
        getStat(): ProcessStat;
        private Stat;
        private _stat;
        processHowEver: boolean;
        static __feilds__(): DProperty<HTMLElement, Controller>[];
        static DPView: DProperty<HTMLElement, Controller>;
        View: HTMLElement;
// ...
    }
}
// ...
    interface Node<T> {
        Depth: number;
        Value: T;
        param?: any;
        children: Node<T>[];
        Parent: Node<T>;
    }
// ...
export declare namespace basic {
    class DomEventHandler<T extends Event, P> implements IEventHandler, EventListenerObject {
        dom: Element;
        event: string;
        private owner;
        private handle;
        private param;
        Started: boolean;
        constructor(dom: Element, event: string, owner: any, handle: (eh: DomEventHandler<T, P>, ev: T, param: P) => void, param?: P);
        Start(): void;
        Pause(): void;
        Dispose(): void;
        Reset(): void;
        handleEvent(evt: Event): void;
        static Dispose(dom: EventTarget, event?: string): void;
    }
}